modifications to type definition
