+++ package.json
@@ -12,10 +12,19 @@
   },
   "version": "0.1.23",
   "description": "Windows Azure Active Directory Client Library for node",
-  "keywords": [ "node", "azure", "AAD", "adal", "adfs", "oauth" ],
+  "keywords": [
+    "node",
+    "azure",
+    "AAD",
+    "adal",
+    "adfs",
+    "oauth"
+  ],
   "main": "./lib/adal.js",
-  "types": "./index.d.ts",
-  "engines": { "node": ">= 0.6.15" },
+  "types": "./lib/adal.d.ts",
+  "engines": {
+    "node": ">= 0.6.15"
+  },
   "dependencies": {
     "@types/node": "^8.0.47",
     "date-utils": "*",
@@ -24,7 +33,7 @@
     "request": ">= 2.52.0",
     "underscore": ">= 1.3.1",
     "xmldom": ">= 0.1.x",
-    "xpath.js": "~1.0.5", 
+    "xpath.js": "~1.0.5",
     "async": ">=0.6.0"
   },
   "devDependencies": {
@@ -32,12 +41,16 @@
     "mocha": "*",
     "nock": "*",
     "sinon": "*",
-    "typescript": "^2.6.1"
+    "typescript": "^2.6.1",
+    "@types/nock": "^8.2.1",
+    "@types/mocha": "^2.2.44",
+    "@types/underscore": "^1.8.4",
+    "@types/sinon": "^2.3.7"
   },
   "scripts": {
     "test": "npm run tsc && mocha -R spec --ui tdd test",
-    "cover" : "./test/util/cover",
-    "doc" : "jsdoc lib",
-    "tsc" : "tsc index.d.ts"
+    "cover": "./test/util/cover",
+    "doc": "jsdoc lib",
+    "tsc": "tsc -p tsconfig.json"
   }
 }
