+++ tsconfig.json
@@ -31,6 +31,7 @@
     "node_modules"
   ],
   "include": [
-    "./index.d.ts"
+    "./lib/adal.d.ts",
+    "./test"
   ]
 }
